feat: Add background scheduler for sending newsletters
This commit adds a new file `runapscheduler.py` which contains the code for setting up a background scheduler using the `BackgroundScheduler` class from the `apscheduler` library. The scheduler is configured to run the `test` function from the `task.py` module at a 10-second interval. The scheduler is started in the `start` function.

The purpose of this change is to automate the sending of newsletters by running the `test` function periodically. This will improve the efficiency and reliability of the newsletter sending process.