feat: Refactor RecommendedJobsView to improve job filtering
The code changes in this commit refactor the `RecommendedJobsView` class in the `newsletter` app to improve the job filtering functionality. Previously, the code was using slicing (`[::10]`) to limit the number of jobs returned, but this has been removed to return all matching jobs. This change ensures that all relevant jobs are included in the response.

This commit addresses the need to display all matching jobs in the newsletter and improves the accuracy of job recommendations for users.