feat: Fix bug in JobAddSerializer
This commit fixes a bug in the JobAddSerializer class in the jobs/serializer.py file. Previously, the condition for updating the instance's attributes was incorrect, resulting in the attributes not being updated when the instance was not edited. This commit corrects the condition to update the instance's attributes regardless of whether it was edited or not.

This change ensures that all validated data is correctly set on the instance, improving the accuracy and reliability of the serializer.

```