feat: Update Authorized view to include is_superuser and is_staff fields in the response
This commit modifies the `Authorized` view in the `users/views.py` file. It adds the `is_superuser` and `is_staff` fields to the response JSON object. This ensures that the client receives information about the user's superuser and staff status.

This change improves the completeness of the response and provides additional information about the user's privileges.