feat: Update website field in Company model to allow longer URLs
The code changes in this commit modify the `website` field in the `Company` model of the `company/models.py` file. Previously, the `website` field had a maximum length of 50 characters, which was insufficient to accommodate longer URLs. This commit updates the `website` field to have a maximum length of 300 characters, allowing for longer URLs to be stored.

This change improves the functionality of the application by enabling the storage of more accurate and complete website URLs for companies.