feat: Update Solr URL in ClearCompany view for deleting company data
The code changes in this commit update the Solr URL in the ClearCompany view of the company/views.py file. Previously, the Solr URL was retrieved from the DATABASE_SOLR environment variable directly. This commit modifies the code to construct the Solr URL by appending "/solr/jobs" to the DATABASE_SOLR environment variable.

This change ensures that the ClearCompany view deletes the company data from the correct Solr collection, improving the accuracy and reliability of the data deletion process.