feat: Update Job model delete method to use Solr query for improved performance
The code changes in this commit modify the delete method of the Job model in the jobs/models.py file. Previously, the method used the requests library to send a POST request to the DATABASE_URL for deleting the job from the database. This approach was inefficient and could result in multiple requests being sent.

This commit improves the performance of the delete method by using the pysolr library to construct and execute a targeted Solr query for deleting the job from the Solr collection. It retrieves the Solr URL from the DATABASE_SOLR environment variable and uses it to create a Solr object. The delete method then uses the Solr object to delete the job from the collection.

This change ensures that the delete method deletes the job from the correct Solr collection, improving the accuracy and efficiency of the deletion process.