feat: Update company sort options in constants.py and add additional fields to CompanySerializer
The code changes in this commit update the COMPANY_SORT_OPTIONS dictionary in the constants.py file of the company module. Two new sort options, "jobs_unpublish_asc" and "jobs_unpublish_desc", are added to sort the companies based on the number of unpublished jobs.

Additionally, the CompanySerializer in the serializers.py file is modified to include two new fields: "published_jobs" and "have_access". The "published_jobs" field calculates the number of published jobs for a company, while the "have_access" field checks if the current user has access to the company.

These changes enhance the functionality of the company module by providing more sorting options and additional information in the CompanySerializer.