feat: Add language field to ScraperAddView and update ScraperModel creation
The code changes in this commit add a new field, "language", to the ScraperAddView class in the scraper/views.py file. This field is used to specify the programming language of the scraper being added.

Additionally, the ScraperModel creation in the ScraperAddView is updated to include the language field. This ensures that the language information is stored in the database for each scraper.

These changes enhance the functionality of the ScraperAddView by allowing users to specify the language of the scraper and store it in the database.