feat: Fix bug in DeleteUser API endpoint
The code changes in this commit fix a bug in the DeleteUser API endpoint in the users/views.py file. Previously, the user_delete object was not being properly deleted when the user is a superuser. This commit removes the commented out code and ensures that the user is deleted correctly.

This change improves the functionality of the DeleteUser API endpoint by fixing the bug that prevented the deletion of users by superusers.